# backend/shortVideoApiService/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /collection:
        get:
            tags:
                - CollectionService
            description: 列出用户的收藏夹
            operationId: CollectionService_ListCollection
            parameters:
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                properties:
                                    code:
                                        type: integer
                                        description: Status code. Zero means success.
                                    msg:
                                        type: string
                                        description: Status message. Could be displayed to user.
                                    data:
                                        $ref: '#/components/schemas/svapi.ListCollectionResponse'
        put:
            tags:
                - CollectionService
            description: 更新收藏夹信息
            operationId: CollectionService_UpdateCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.UpdateCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                properties:
                                    code:
                                        type: integer
                                        description: Status code. Zero means success.
                                    msg:
                                        type: string
                                        description: Status message. Could be displayed to user.
                                    data:
                                        $ref: '#/components/schemas/svapi.UpdateCollectionResponse'
        post:
            tags:
                - CollectionService
            description: 创建收藏夹
            operationId: CollectionService_CreateCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.CreateCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                properties:
                                    code:
                                        type: integer
                                        description: Status code. Zero means success.
                                    msg:
                                        type: string
                                        description: Status message. Could be displayed to user.
                                    data:
                                        $ref: '#/components/schemas/svapi.CreateCollectionResponse'
        delete:
            tags:
                - CollectionService
            description: 删除收藏夹
            operationId: CollectionService_RemoveCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.RemoveCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                properties:
                                    code:
                                        type: integer
                                        description: Status code. Zero means success.
                                    msg:
                                        type: string
                                        description: Status message. Could be displayed to user.
                                    data:
                                        $ref: '#/components/schemas/svapi.RemoveCollectionResponse'
    /collection/video:
        get:
            tags:
                - CollectionService
            description: 列出收藏夹中的视频
            operationId: CollectionService_ListVideo4Collection
            parameters:
                - name: collectionId
                  in: query
                  description: '@gotags: json:"collectionId,omitempty,string"'
                  schema:
                    type: string
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                properties:
                                    code:
                                        type: integer
                                        description: Status code. Zero means success.
                                    msg:
                                        type: string
                                        description: Status message. Could be displayed to user.
                                    data:
                                        $ref: '#/components/schemas/svapi.ListVideo4CollectionResponse'
        post:
            tags:
                - CollectionService
            description: 将视频添加到收藏夹
            operationId: CollectionService_AddVideo2Collection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.AddVideo2CollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                properties:
                                    code:
                                        type: integer
                                        description: Status code. Zero means success.
                                    msg:
                                        type: string
                                        description: Status message. Could be displayed to user.
                                    data:
                                        $ref: '#/components/schemas/svapi.AddVideo2CollectionResponse'
        delete:
            tags:
                - CollectionService
            description: 从收藏夹中移除视频
            operationId: CollectionService_RemoveVideoFromCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.RemoveVideoFromCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                properties:
                                    code:
                                        type: integer
                                        description: Status code. Zero means success.
                                    msg:
                                        type: string
                                        description: Status message. Could be displayed to user.
                                    data:
                                        $ref: '#/components/schemas/svapi.RemoveVideoFromCollectionResponse'
...
components:
    schemas:
...
        svapi.AddVideo2CollectionRequest:
            type: object
            properties:
                collectionId:
                    type: string
                    description: '@gotags: json:"collectionId,omitempty,string"'
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
        svapi.AddVideo2CollectionResponse:
            type: object
            properties: {}
...
        svapi.Collection:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                userId:
                    type: string
                    description: '@gotags: json:"userId,omitempty,string"'
                name:
                    type: string
                description:
                    type: string
        svapi.CollectionVideo:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
                coverUrl:
                    type: string
                title:
                    type: string
                description:
                    type: string
                videoUrl:
                    type: string
...
        svapi.CreateCollectionRequest:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
        svapi.CreateCollectionResponse:
            type: object
            properties:
                collection:
                    $ref: '#/components/schemas/svapi.Collection'
...
        svapi.ListCollectionResponse:
            type: object
            properties:
                collections:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.Collection'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationResponse'
...
        svapi.ListVideo4CollectionResponse:
            type: object
            properties:
                videos:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.CollectionVideo'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationResponse'
...
        svapi.PaginationResponse:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                count:
                    type: integer
                    format: int32
...
        svapi.RemoveCollectionRequest:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
        svapi.RemoveCollectionResponse:
            type: object
            properties: {}
...
        svapi.RemoveVideoFromCollectionRequest:
            type: object
            properties:
                collectionId:
                    type: string
                    description: '@gotags: json:"collectionId,omitempty,string"'
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
        svapi.RemoveVideoFromCollectionResponse:
            type: object
            properties: {}
...
        svapi.UpdateCollectionRequest:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                description:
                    type: string
        svapi.UpdateCollectionResponse:
            type: object
            properties:
                collection:
                    $ref: '#/components/schemas/svapi.Collection'
...